+++ mobile/android/android-components/components/browser/engine-gecko-nightly/src/androidTest/java/mozilla/components/browser/engine/gecko/fetch/geckoview/GeckoViewFetchTestCases.kt
@@ -109,4 +109,10 @@ class GeckoViewFetchTestCases : mozilla.components.tooling.fetch.tests.FetchTest
     override fun put201FileUpload() {
         super.put201FileUpload()
     }
+
+    @Test
+    @UiThreadTest
+    override fun get200WithCookiePolicy() {
+        super.get200WithCookiePolicy()
+    }
 }