+++ mobile/android/android-components/components/lib/fetch-httpurlconnection/src/main/java/mozilla/components/lib/fetch/httpurlconnection/HttpURLConnectionClient.kt
@@ -9,9 +9,12 @@ import mozilla.components.concept.fetch.Headers
 import mozilla.components.concept.fetch.MutableHeaders
 import mozilla.components.concept.fetch.Request
 import mozilla.components.concept.fetch.Response
+import mozilla.components.lib.fetch.httpurlconnection.HttpURLConnectionClient.Companion.getOrCreateCookieManager
 import java.io.FileNotFoundException
 import java.io.IOException
 import java.io.InputStream
+import java.net.CookieHandler
+import java.net.CookieManager
 import java.net.HttpURLConnection
 import java.net.URL
 import java.util.zip.GZIPInputStream
@@ -29,6 +32,15 @@ class HttpURLConnectionClient : Client() {
         connection.addBodyFrom(request)
 
         return connection.toResponse()
+    }
+
+    companion object {
+        fun getOrCreateCookieManager(): CookieManager {
+            if (CookieHandler.getDefault() == null) {
+                CookieHandler.setDefault(CookieManager())
+            }
+            return CookieHandler.getDefault() as CookieManager
+        }
     }
 }
 
@@ -61,6 +73,17 @@ private fun HttpURLConnection.setupWith(request: Request) {
 
     request.readTimeout?.let { (timeout, unit) ->
         readTimeout = unit.toMillis(timeout).toInt()
+    }
+
+    // HttpURLConnection can't be configured to omit cookies. As
+    // a workaround, we delete all cookies we have stored for
+    // the request URI.
+    val cookieManager = getOrCreateCookieManager()
+    if (request.cookiePolicy == Request.CookiePolicy.OMIT) {
+        val uri = URL(request.url).toURI()
+        for (cookie in cookieManager.cookieStore.get(uri)) {
+            cookieManager.cookieStore.remove(uri, cookie)
+        }
     }
 }